ito : add swagger conf + projectResource operation API Desc

+++ src/main/java/org/ipc/synapsis/projet/resource/ProjetResource.java
@@ -10,6 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.core.MediaType;
 import java.net.URI;
@@ -25,42 +29,60 @@ public class ProjetResource  {
     @Autowired
     IProjetService projetService;
 
-
+    @ApiOperation(value = "ADD a new Projet")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 201, message = "Projet creation : Done", response = URI.class) })
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON, produces = MediaType.APPLICATION_JSON )
     public ResponseEntity add(@RequestBody final ProjetIn projetIn,HttpServletRequest request) {
         LOGGER.debug("Start call of the web service add new 'Projet',{}",projetIn);
         UUID id = projetService.add(projetIn);
         LOGGER.debug("End call of the web service add new 'Projet',{}",projetIn);
-
         return ResponseEntity.created(URI.create(request.getRequestURL().append("/"+id).toString())).build();
-
     }
 
-
+    @ApiOperation(value = "Update a Projet")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projet Updated : Done", response = URI.class),
+			@ApiResponse(code = 404, message = "Projet Updated : Not Found", response = URI.class) })
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON, produces = MediaType.APPLICATION_JSON )
     public ResponseEntity update(@PathVariable final String id, @RequestBody final ProjetIn projetIn) {
         LOGGER.debug("Start call of the web service update 'Projet',{}", projetIn);
-        projetService.update(id, projetIn);
-        LOGGER.debug("End call of the web service update 'Projet',{}", projetIn);
-        return ResponseEntity.ok().build();
-
+        if(projetService.isHere(id))
+        {
+	        projetService.update(id, projetIn);
+	        LOGGER.debug("End call of the web service update 'Projet',{}", projetIn);
+	        return ResponseEntity.ok().build();
+        }
+        else return ResponseEntity.notFound().build();
     }
 
+    @ApiOperation(value = "Get a Projet")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projet found", response = URI.class),
+			@ApiResponse(code = 404, message = "Projet not found", response = URI.class),
+			@ApiResponse(code = 422, message = "Projet : technical Error", response = URI.class) })
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ResponseEntity get(@PathVariable final String id) throws Exception {
         LOGGER.debug("Start call of the web service get 'Projet' by id, id={}",id);
-        ProjetOut projetOut = null;
-        try {
-            projetOut = projetService.get(id);
-        } catch (Exception e) {
-            LOGGER.warn("Exception get Projet");
-
+        if(projetService.isHere(id))
+        {
+	        ProjetOut projetOut = null;
+	        try {
+	            projetOut = projetService.get(id);
+	        } catch (Exception e) {
+	            LOGGER.warn("Exception get Projet");
+	            return ResponseEntity.unprocessableEntity().build();
+	        }
+	        LOGGER.debug("End call of  the web service get 'Projet' by id, id={}",id);
+	        return ResponseEntity.ok(projetOut);
         }
-        LOGGER.debug("End call of  the web service get 'Projet' by id, id={}",id);
-        return ResponseEntity.ok(projetOut);
+        else return ResponseEntity.notFound().build();
     }
 
-
+    @ApiOperation(value = "Get all Projet")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projets found", response = URI.class),
+			@ApiResponse(code = 400, message = "Projet : technical Error", response = URI.class) })
     @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON )
     public ResponseEntity getAll() {
         LOGGER.debug("Start call of the web service get all 'Projet'");
@@ -69,20 +91,31 @@ public class ProjetResource  {
         return ResponseEntity.ok(projetOutList);
     }
 
-
-    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, consumes = MediaType.APPLICATION_JSON)
+    @ApiOperation(value = "Delete a Projet")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 204, message = "Projet deleted", response = URI.class),
+			@ApiResponse(code = 404, message = "Projet not found", response = URI.class) })
+    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON)
     public ResponseEntity remove(@PathVariable final String id) throws Exception {
         LOGGER.debug("Start call of the web service delete 'Projet' by id,id={}",id);
-        try {
-            projetService.remove(id);
-        }catch (Exception e) {
-            LOGGER.error("Resource layer Cannot parse Sting to UUID");
-
+        if(projetService.isHere(id))
+        {
+        	try {
+	            projetService.remove(id);
+		        LOGGER.debug("End call of the web service delete 'Projet' by id,id={}",id);
+		        return ResponseEntity.ok().build();
+	        }catch (Exception e) {
+	            LOGGER.error("Resource layer Cannot parse Sting to UUID");
+	            return ResponseEntity.badRequest().build();
+	        }
         }
-        LOGGER.debug("End call of the web service delete 'Projet' by id,id={}",id);
-        return ResponseEntity.ok().build();
+        else return ResponseEntity.notFound().build();
+        
     }
 
+    @ApiOperation(value = "Find a projet by Title")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projet Found", response = URI.class) })
     @RequestMapping(value = "/title/{title}", method = RequestMethod.GET)
     public ResponseEntity findByTitle(@PathVariable final String title) throws Exception {
         LOGGER.debug("Start call of the web service get 'Projet' by , title={}",title);
@@ -97,7 +130,10 @@ public class ProjetResource  {
         return ResponseEntity.ok(listProjetOut);
     }
 
-    @RequestMapping(value = "/criterias", method = RequestMethod.GET)
+    @ApiOperation(value = "Find a projet by Title containing..")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projet Found", response = URI.class) })
+    @RequestMapping(value = "/criterias/title", method = RequestMethod.GET)
     public ResponseEntity findByTitleContaining(@RequestParam final String title) throws Exception {
         LOGGER.debug("Start call of the web service get 'Projet' by , title={}",title);
         List<ProjetOut> listProjetOut = null;
@@ -108,6 +144,23 @@ public class ProjetResource  {
 
         }
         LOGGER.debug("End call of  the web service get 'Projet' by , title={}",title);
+        return ResponseEntity.ok(listProjetOut);
+    }
+    
+    @ApiOperation(value = "Find a projet by Description containing..")
+	@ApiResponses(value = { 
+			@ApiResponse(code = 200, message = "Projet Found", response = URI.class)})    
+    @RequestMapping(value = "/criterias/description", method = RequestMethod.GET)
+    public ResponseEntity findByDescriptionContaining(@RequestParam final String description) throws Exception {
+        LOGGER.debug("Start call of the web service get 'Projet' by , description={}",description);
+        List<ProjetOut> listProjetOut = null;
+        try {
+            listProjetOut = projetService.findByDescriptionContaining(description);
+        } catch (Exception e) {
+            LOGGER.warn("Exception get Projet");
+
+        }
+        LOGGER.debug("End call of  the web service get 'Projet' by , description={}",description);
         return ResponseEntity.ok(listProjetOut);
     }
